+++ CHANGELOG.md
@@ -24,10 +24,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Trades`
   * `all/0` and `all/1` for listing trades in the system
   * `all_for_order_book/2` and `all_for_order_book/3` for listing trades for an order book
-- `Order Book`
+- `OrderBook`
   * `get/1` for getting order book details
 - `PaymentPaths`
-  * `all/0` and `all/1` for listing payment paths
+  * `all/3` and `all/4` for listing payment paths
+- `TradeAggregations`
+  * `all/2` and `all/3` for listing trade aggregations
 
 ## [0.1.1] - 2018-01-27
 ### Fixed